requirements.txt
# Vector Databases & Similarity Search
faiss-cpu

# OpenAI
openai>=1.10.0
tiktoken>=0.5.2

# UI Framework
gradio==4.7.1
setuptools>=65.5.1

# Data Processing & Utils
numpy==1.26.2
pandas==2.1.3

# Environment Variables
python-dotenv==1.0.0
